Extract binary PLY data handlers into engine/src/binaryDataHandlers.ts
Moves loadWithPerf, handleUltimateRawBinaryUri, handleUltimateRawBinaryData
(the hot-path binary PLY parser), handleDirectTypedArrayData, and
handleBinarySpatialData out of main.ts as thin wrappers. The hot loops
only ever touched local variables, so this is a pure code-motion change
with no behavioral or performance impact.
